rename from seeds -> inventory
