+++ e2e/pages/inventory/create.py
