+++ e2e/pages/inventory/edit.py
