+++ e2e/steps/common_steps.py
@@ -8,7 +8,7 @@ from e2e.pages.login import LoginPage
 from e2e.pages.maps.create import MapCreatePage
 from e2e.pages.maps.management import MapManagementPage
 from e2e.pages.maps.planting import MapPlantingPage
-from e2e.pages.seeds.management import SeedManagementPage
+from e2e.pages.inventory.management import SeedManagementPage
 
 
 @given("I am on the seed management page")
@@ -23,7 +23,7 @@ def smp(page: Page) -> SeedManagementPage:
     hp = HomePage(page)
     lp = LoginPage(page)
     login(hp, lp)
-    smp = hp.to_seed_management_page()
+    smp = hp.to_inventory_page()
     smp.verify()
     return smp
